fixed: wrong offset in last commit. forgot height of UV textures is half. ie offset had to be 1/8th pixel instead of 1/4th.
This is needed as when an image is split into two fields, so is the chroma channel..
first line of chroma in first field is then seen to be closer to first line of luminance than second.
in second field first line of chroma is closer to second line than first.