a bit of code cleanup (extra checks for pointers + removed some .h file dependencies)
dvdplayer: some dvds used the wrong subtitle colors that resulted in funny subtitles. Thx to frov for sending me a dvd that had this problem
some subtitles still have color issues when selecting only the .vob file to play

dvdplayer: changed the way the number of subtitles and audiostreams are found.
hopefully it will fix a lot of audio / subtitle issues in the osd (unknown streams and such)

dvdplayer: some debugging utilities