changed: force calls to mplayer to align the stack to 16 byte boundary on function call. added: helper defines to make aligned function calls using dllloader
GCC compiled dlls assumes stack is aligned on function entry, it will only ensure that the alignment is kept when going deeper, but this will be wrong if the caller is non gcc based. This misalignment will result in worse performance for mmx based rutines that uses variables on stack, and segfaults for sse based ones (no segfaults in current mplayer/libavcodec as I know of). Thus this change may increase performance for codecs wich are heavly mmx optimized.