added: Basic PS3 Sixaxis support as an event client. Wave it around with the PS button pressed and watch is act like a mouse.
added: A simple Bluetooth HID server to allow the PS3 Blu-Ray remote
and Sixaxis to auto connect.

NOTE 1: The HID server needs other HID servers to be disabled (including Bluez's
if it has been enabled in /etc/default/bluetooth and the bluetooth input
server in Bluetooth system preferences)

NOTE 2: The HID server requires root access currently because it listens on 
Bluetooth L2CAP PSMs 0x11 and 0x13, both of which are privlieged ports.