fixed: thin line to the right and/or left of gui image elements.
Since we sometimes use only a part of power of two textures to store image data. GL's clamp_to_edge never works, as we never touch the edge. Instead it samples unwritten data outside the actual image when it renders the texture. This simulates clamp_to_edge by filling the pixels outside the actual image in the texture with the last valid pixel.