Ticket #8661: xbmc-glxvisuals.patch
Currently, XBMC checks the GLX version. If it is >= 1.3, fbConfigs are requested and the first matching one with depth = 24 is chosen. If the GLX version < 1.3, it just relies on glXChooseVisual instead. There are some problems with this approach. First, glXChooseVisual may choose a visual that has depth different from the SDL window, e.g. 32 (while window = 24) as happened on my system (nouveau+dri_swrast, of course other systems can well be affected). This causes an X Bad Match protocol error later. However, just selecting a visual matching the window depth may not be enough, due to a second issue: glXMakeCurrent documentation says that Bad Match is generated if drawable (i.e. SDL window) was created with a different visual than ctx (for which we find visual ourself). And indeed I encountered this with libGL in forced indirect mode (again, may affect other implementations). This patch rewrites the visual selection code; now it first tries to visual of the SDL window. If that visual does not meet the requirements, we search for an acceptable visual with same depth as the SDL window (some implementations, including proprietary NVIDIA, seem to allow using a different visual), this time with calls that are supported by GLX < 1.3 as well.